W-11362986 (refactor): Removed FileAssertionTest and use amf-core class instead
