+++ amf-cli/jvm/src/test/scala/amf/cli/internal/tools/ModelExporterTest.scala
@@ -1,6 +1,6 @@
 package amf.cli.internal.tools
 
-import amf.io.FileAssertionTest
+import amf.core.io.FileAssertionTest
 import org.scalatest.funsuite.AsyncFunSuite
 
 class ModelExporterTest extends AsyncFunSuite with FileAssertionTest {
