+++ amf-cli/shared/src/test/scala/amf/cli/internal/convert/WrapperTests.scala
@@ -36,7 +36,8 @@ import amf.core.internal.convert.CoreClientConverters.{ClientFuture, _}
 import amf.core.internal.remote.Mimes._
 import amf.core.internal.remote._
 import amf.core.internal.resource.{ClientResourceLoaderAdapter, StringResourceLoader}
-import amf.io.{FileAssertionTest, MultiJsonldAsyncFunSuite}
+import amf.core.io.FileAssertionTest
+import amf.io.MultiJsonldAsyncFunSuite
 import amf.shapes.client.platform.ShapesConfiguration
 import amf.shapes.client.platform.model.domain.{AnyShape, NodeShape, ScalarShape, SchemaShape}
 import amf.shapes.client.platform.render.JsonSchemaShapeRenderer
