+++ amf-cli/shared/src/test/scala/amf/event/AMFEventReportBuilderTest.scala
@@ -3,7 +3,7 @@ package amf.event
 import amf.core.client.scala.config.event.{AMFEventReportBuilder, TimedEvent}
 import amf.core.client.scala.config.{AMFEvent, GroupedEvent}
 import amf.core.internal.unsafe.PlatformSecrets
-import amf.io.FileAssertionTest
+import amf.core.io.FileAssertionTest
 import org.mulesoft.common.test.AsyncBeforeAndAfterEach
 import org.scalatest.matchers.should.Matchers
